Fine Motor Control Fixes uhhhhhhgggghhhh Car can now accuratly travel between all floors at default max speed and acel. Changing those values screws things up though.

+++ include/ServiceQueueControlTask.h
@@ -5,11 +5,11 @@
 #include "CommonIncludes.h"
 #include "TypesAndGlobalVars.h"
 #include "UartDriver.h"
-#include <math.h> ///// Uuuuhhggggggg
+
 
 void ServiceQueueControlTask(void *param_struct);
-MotorMessage CreateNewMotorMessage(int current_max_speed,
-                                            int current_acel,
+MotorMessage CreateNewMotorMessage(float current_max_speed,
+                                            float current_acel,
                                             int current_floor,
                                             int requested_floor);
 #endif	/* SERVICEQUEUECONTROLTASK_H */
